py-middleware/requirements.txt
eventlet
Flask-SocketIO
RPi.GPIO
Pillow
